UTCL-MFG-DATA_FOUNDATION_PLATEFORM/CBM-Functions/plants_data_not_coming_alerts/requirements.txt
pyodbc
requests
pandas
numpy
boto3
python-dotenv

